+++ package.json
@@ -55,5 +55,4 @@
       "last 1 safari version"
     ]
   }
-
 }
